fixed bug on treasury total pages validation
